Get text input to be plumbed through to the RenderWidgetHostViewAura instance on Windows in desktop mode.
This is achieved by passing the InputMethod instance to the InputMethodWin class which handles text input.
In the OnChar handling we look for the TextInputClient implementation on the focused view and on failing to
find that we retrieve that from the InputMethod instance.

BUG=157172
R=ben
Review URL: https://codereview.chromium.org/11228032

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@163446 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 35cd96d167a100094727640b7832d9d3a2b3eb06