Store the dimensions of the first finger's enclosing rectangle when creating an ET_GESTURE_TWO_FINGER_TAP event
We would like to be able to perform touch target fuzzing on the first
finger of a two-finger tap gesture in the same way as a long press gesture.
This CL stores the dimensions of the first finger's enclosing rectangle
when creating a two-finger tap gesture so that this information can be
available to WebKit. The previously-unused |delta_x| and |delta_y| values
in GestureEventDetails are now used in plumbing the enclosing rectangle's
width and height values.

This depends on https://bugs.webkit.org/show_bug.cgi?id=99763 landing
and being gardened in first.

BUG=140024

Review URL: https://chromiumcodereview.appspot.com/11232033

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@163575 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a3a7850328fda48aff5743aacba31dcf506b7953