views: Start converting View into an EventTarget.
The EventDispatcher interface is not acually used for dispatching events yet.
That will happen once RootView is turned into an EventDispatcher. But this first
step prepares scroll, touch, and gesture event dispatching in views so that it
is compatible with the EventDispatcher interface.

BUG=none

Review URL: https://codereview.chromium.org/11364015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@165288 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3630942bd18a8f06d67412b737340a2eaa845753