ui: Use Vector2d in the View::CalculateOffsetToAncestorWithLayer API
This function computes an offset, so it should be a vector. Since we have
math operations now, instead of passing in a Point*, we return a new
Vector2d, which we can add to other Vector2ds, or use as is.

Depends on https://codereview.chromium.org/11293194/

R=pkasting
BUG=158615

Review URL: https://codereview.chromium.org/11362179

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@167285 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3f1e150ba3727dfecf666e9a6752ac9942f85cd4