Make views::Link underlining configurable
This CL allows the underlining of a views::Link to be enabled and
disabled. The functionality will be needed for the public account user
card. As per its UI design, the user card contains a link that should not
be underlined.

BUG=152928

Review URL: https://chromiumcodereview.appspot.com/11412010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@168226 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b40a1272637611b8a091b11c7ec84611b5e01d66