Fix the html select tag showing up at the wrong position. This was a regression from r166446.
The fix is to watch for the top level aura::Window's movement. In Ash, this would be an aura::Window that is the parent of the WebContents. In desktop Aura, this would be the aura::RootWindow. The change watches for this in WebContentsViewAura, similar to how WebContentsViewWin does this with PositionChangedMessageFilter. The reason is that the parent being watched needs to be updated with tab drags, so it's more convenient to do this at WebContentsViewAura's parent, instead of doing it at the RenderWidgetHostViewAura level (since that would need a list of parents).

This change also fixes OnHostMoved not being called on Windows.

In the reland of this patch, I've also fixed WebContentsView implementations holding on to deleted RenderWidgetHostView objects after an interstitial is hidden.

BUG=161174
Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=168558
Review URL: https://codereview.chromium.org/11308083

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@168767 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8dbb764ef23f647729fb9a538229996d5e504aad