Support creating virtual contexts.
This creates virtual contexts within a GLShareGroup, when
'--enable-virtual-gl-contexts' is given on the command line.
It requires the shared real GL context, which is created
from the default offscreen surface configuration, to be compatible
with any surface that is made current with a virtual context.

BUG=160710


Review URL: https://chromiumcodereview.appspot.com/11308151

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@170148 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ee1e6aac01990666814653f6d14be4c048e792aa