touch: Fix ctrl+taps for buttons on bookmark bar and toolbar.
Ctrl+Tapping on the buttons in bookmark bar/toolbar should open a new tab. But
since the callbacks looked only at the event-flags for the mouse-events, it
didn't work correctly for taps. This is a fix for that.

BUG=163377

Review URL: https://codereview.chromium.org/11419227

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@170238 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 437e1bbc3ec40d18913c7a698515995ce1754f10