Fix the scale pivot for views corewm hide window animations.
Right now, we use 0.025 * bounds as the pivot, which basically amounts to the
upper left corner of the window. I don't think that was intended. I think that
we mean to scale about the center of the window when hiding.

BUG=163385


Review URL: https://chromiumcodereview.appspot.com/11411283

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@170515 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b949fcb52666d423b7c66d9a7a69d482bf056513