Remove ui::LayerAnimationSequence::duration and ui::LayerAnimationElement::duration.
This change is in preparation for threading certain ui animations. When we do
this, the actual duration of an element will not be known ahead of time, since 
dispatching the element to the compositor thread will involve a small delay.

BUG=164206

Review URL: https://chromiumcodereview.appspot.com/11316315

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@171112 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4a386e484f982655856fb05e43fd5d03a6b237c7