Remove calls to StoreFocusedView and RestoreFocusedView in DesktopNativeWidegetAura::OnBlur/OnActivated. This is the wrong place since OnBlur/OnActivated is invoked when the window associated with the DNWA loses focus/activation to a child window. At this time we don't want to store/restore focus. Instead the right place to do this is in DesktopRootWindowHostWin on callbacks from HWNDMessageHandler. The code is already there, so this code can be removed entirely.
This code is causing problems when you click on a child window (such
as the renderer). When this happens DNWA gets OnBlur so that if it
tries to store focus it conflicts with the renderer trying to set
itself as the focus target.

BUG=none
TEST=none
R=ben@chromium.org


Review URL: https://chromiumcodereview.appspot.com/11471002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@171578 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 77cf95b5898337609b294c2d004451d7db732890