Desktop aura: Start removing views::Widget::CreateWindow()
We have lots of code in chrome which just assumes that it can throw up a dialog
with no context. This will not work correctly in ash/aura combo builds. This is
part 1 of threading context through chrome so each dialog will be displayed
on the correct desktop, which is a subgoal of removing the global
StackingClient.

Widget::CreateWindow() and Widget::CreateWindowWithBounds() will be
transitioned to Widget::CreateWindowWithContext() and
Widget::CreateWindowWithContextAndBounds().

BUG=161882

Review URL: https://codereview.chromium.org/11519022

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@172650 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6e3920611d6243e2860235146baa11c49a49f668