cc: Add point-based UV coordinate on TextureLayer
Previously, the uv coordinates was using gfx::RectF as a struct.
This was limitating because it could not handle negative width
and heights. Soring them as 2 gfx::PointF for the Top-Left and
Bottom-Right points aleviate that restriction.
The current implementation of the phone UI rely on this feature.

BUG=168927

Review URL: https://chromiumcodereview.appspot.com/11783094

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@177288 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e692f68f27275e4573c4cf8ec23a731dd3e33ffd